Sales is toegevoegd.
Als salesmedewerker kan ik nu bedrijven toevoegen , bewerken, verwijderen.
En offertes sturen op basis van de geselecteerde producten uit het systeem.
Daarnaast kan ik alle verzonden offertes bekijken en bewerken, zolang ze nog niet zijn verstuurd. Ook kan ik notities toevoegen, bewerken en verwijderen. De salesmedewerker heeft nu toegang tot de onderhoudsafdeling

+++ app/Livewire/Purchase/ProductManager.php
@@ -144,13 +144,19 @@ class ProductManager extends Component
             'categories' => ProductCategory::all(),
         ]);
     }
-
     public function deleteProduct($id)
     {
-        $product = Product::findOrFail($id);
+        $product = Product::withCount('inventoryMovements')->findOrFail($id);
 
-        $product->delete();
+        if ($product->inventory_movements_count > 0) {
+            session()->flash(
+                'error',
+                'Dit product is al gebruikt/besteld en kan niet worden verwijderd.'
+            );
+            return;
+        }
 
+        $product->delete(); // soft delete
         session()->flash('success', 'Product verwijderd.');
     }
